Finetune initial state
