+++ src/state/index.ts
@@ -1,3 +1,2 @@
-export * from './actions';
-export * from './reducer';
-export * from './state';
+export * from "./reducer";
+export * from "./state";
