+++ src/types.ts
@@ -4,21 +4,6 @@ export interface Diagnose {
   latin?: string;
 }
 
-export interface Entry {
-  id: string;
-  description: string;
-  date: string;
-  specialist: string;
-  diagnoseCodes?: Array<Diagnose["code"]>;
-}
-
-export enum HealthCheckRating {
-  "Healthy" = 0,
-  "LowRisk" = 1,
-  "HighRisk" = 2,
-  "DiagnosedCondition" = 3
-}
-
 export enum Gender {
   Male = "male",
   Female = "female",
@@ -32,5 +17,4 @@ export interface Patient {
   dateOfBirth: string;
   occupation: string;
   gender: Gender;
-  entries: Entry[];
 }
